1309-file-size-validation

+++ src/user/UserService.js
@@ -76,10 +76,12 @@ const getUser = async (id) => {
 const updateUser = async (id, updatedBody) => {
   const user = await User.findOne({ where: { id: id } });
   user.username = updatedBody.username;
-  if (user.image) {
-    await FileService.deleteProfileImage(user.image);
+  if (updatedBody.image) {
+    if (user.image) {
+      await FileService.deleteProfileImage(user.image);
+    }
+    user.image = await FileService.saveProfileImage(updatedBody.image);
   }
-  user.image = await FileService.saveProfileImage(updatedBody.image);
   await user.save();
   return {
     id: id,
